Unify route import rewrite mechanics without changing loaders
Route module loading now shares leaf package and rewrite helpers behind the private import rewrite adapter while preserving the route-loader facade and runtime-specific orchestration. Discovery remains local because its cache and fallback policies differ from route loading.

Constraint: Keep external-import-rewriter.ts as the stable facade and runtime branch orchestrator.

Constraint: Preserve Task 3 route package-entry join-plus-containment behavior and subpath containment handling.

Rejected: Move rewriteExternalImports and Node route orchestration | preflight keeps byte-sensitive branch order and Veryfront export-map handling local.

Rejected: Share discovery rewrite entrypoints | discovery fallback and cache semantics differ from route loader semantics.

Confidence: high

Scope-risk: moderate

Tested: RED deno test --no-check --allow-all src/transforms/import-rewriter/core.test.ts failed with missing route-adapter.ts

Tested: deno test --no-check --allow-all src/transforms/import-rewriter/core.test.ts src/discovery/import-rewriter.test.ts src/discovery/transpiler.test.ts src/routing/api/module-loader/external-import-rewriter.test.ts src/routing/api/module-loader/loader.test.ts src/routing/api/module-loader/loader-helpers.test.ts

Tested: deno test --no-check --allow-all src/routing/api/module-loader/loader.test.ts

Tested: deno check src/transforms/import-rewriter/route-adapter.ts src/routing/api/module-loader/external-import-rewriter.ts src/transforms/import-rewriter/core.test.ts src/routing/api/module-loader/loader.test.ts src/routing/api/module-loader/loader-helpers.test.ts

Tested: deno fmt --check src/transforms/import-rewriter/route-adapter.ts src/routing/api/module-loader/external-import-rewriter.ts src/transforms/import-rewriter/core.test.ts src/routing/api/module-loader/loader.test.ts src/routing/api/module-loader/loader-helpers.test.ts

Tested: deno lint src/transforms/import-rewriter/route-adapter.ts src/routing/api/module-loader/external-import-rewriter.ts src/transforms/import-rewriter/core.test.ts src/routing/api/module-loader/loader.test.ts src/routing/api/module-loader/loader-helpers.test.ts

Tested: git diff --check

Not-tested: Full repository test suite
